requirements.txt
streamlit
firebase-admin
pandas
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
